Api3/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
sqlalchemy==2.0.41
pydantic==2.11.5
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
google-generativeai==0.3.2
pydantic-settings==2.1.0
